Extract RBF solver to a separate class
