+++ src/utils/EigenHelperFunctions.cpp
@@ -55,27 +55,5 @@ void append(
   v(n) = value;
 }
 
-Eigen::VectorXd reduceVector(
-    const Eigen::VectorXd &  fullVector,
-    const std::vector<bool> &deadAxis)
-{
-  int deadDimensions = 0;
-  int dimensions     = deadAxis.size();
-  for (int d = 0; d < dimensions; d++) {
-    if (deadAxis[d])
-      deadDimensions += 1;
-  }
-  PRECICE_ASSERT(dimensions > deadDimensions, dimensions, deadDimensions);
-  Eigen::VectorXd reducedVector(dimensions - deadDimensions);
-  int             k = 0;
-  for (int d = 0; d < dimensions; d++) {
-    if (not deadAxis[d]) {
-      reducedVector[k] = fullVector[d];
-      k++;
-    }
-  }
-  return reducedVector;
-}
-
 } // namespace utils
 } // namespace precice
