+++ src/utils/EigenHelperFunctions.hpp
@@ -15,9 +15,6 @@ void appendFront(Eigen::MatrixXd &A, Eigen::VectorXd &v);
 
 void removeColumnFromMatrix(Eigen::MatrixXd &A, int col);
 
-/// Deletes all dead directions from fullVector and returns a vector of reduced dimensionality.
-Eigen::VectorXd reduceVector(const Eigen::VectorXd &fullVector, const std::vector<bool> &deadAxis);
-
 void append(Eigen::VectorXd &v, double value);
 
 template <typename Derived1>
